refactor: components page for showcase
